requirements.txt
flask
maxgram
requests
psycopg2-binary
